+++ src/listpack.h
@@ -74,7 +74,7 @@ unsigned char *lpLast(unsigned char *lp);
 unsigned char *lpNext(unsigned char *lp, unsigned char *p);
 unsigned char *lpPrev(unsigned char *lp, unsigned char *p);
 uint32_t lpBytes(unsigned char *lp);
-unsigned char *lpSeek(unsigned char *lp, int index);
+unsigned char *lpSeek(unsigned char *lp, long index);
 void lpRepr(unsigned char *lp);
 typedef int (*listpackValidateEntryCB)(unsigned char *p, void *userdata);
 int lpValidateIntegrity(unsigned char *lp, size_t size, int deep,
